Agents panel: every Save click is an attempt; footer, line and tint clear together
Final review gate on 1f084fa1 (codex sol, claude-fable-5, own fable-5.1 lane,
audit): the Every-N cadence check returned before noteSubagentsSaveAttempt(),
so a Save aborted by a malformed cadence left an unrouted card neutral —
the stamp now precedes every draft validation. The footer status that names a
roster error after Save is cleared by the same fix that clears the card's tint
and the section line, so no surface keeps reporting a stale draft.

The Agents-panel acceptance moves to tests/test_ui_smoke_agents_panel.py — a
sibling of the main smoke module (which sits at its byte gate), reusing its
server fixture through pytest_plugins as the other UI smokes do — and gains:
typing into a SAVED card reads "Draft ·" in place (the old step ran after
Add, which already repaints as Draft, so it pinned nothing), the cadence-
aborted Save still judges the roster, the footer clears with the fix.
Static pin: the stamp precedes the cadence check. ROUTE_HINT is module-
private; .reviewer-slot-rows uses the space token.
