+++ web/style.css
@@ -6398,7 +6398,7 @@ textarea.chat-input {
     font-weight: 600;
     color: var(--text-primary);
 }
-.reviewer-slot-rows { display: flex; flex-direction: column; gap: 8px; }
+.reviewer-slot-rows { display: flex; flex-direction: column; gap: var(--space-2); }
 .reviewer-slot-row {
     border: 1px solid var(--divider);
     border-radius: 8px;
